requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
email-validator==2.1.1
